move test_is_cmt_option to test_cmt and changed default hit_threshold to cmt style
