+++ straxen/plugins/peaklet_processing.py
@@ -6,7 +6,7 @@ import straxen
 from .pulse_processing import HITFINDER_OPTIONS, HITFINDER_OPTIONS_he, HE_PREAMBLE
 from strax.processing.general import _touching_windows
 
-from straxen.get_corrections import is_cmt_option, test_is_cmt_option
+from straxen.get_corrections import is_cmt_option
 
 export, __all__ = strax.exporter()
 
@@ -113,7 +113,6 @@ class Peaklets(strax.Plugin):
         # Check config of `hit_min_amplitude` and define hit thresholds
         # if cmt config
         if is_cmt_option(self.config['hit_min_amplitude']):
-            test_is_cmt_option()
             self.hit_thresholds = straxen.get_correction_from_cmt(self.run_id,
                 self.config['hit_min_amplitude'])
         # if hitfinder_thresholds config
@@ -443,7 +442,6 @@ class PeakletsHighEnergy(Peaklets):
         # Check config of `hit_min_amplitude_he` and define hit thresholds
         # if cmt config
         if is_cmt_option(self.config['hit_min_amplitude_he']):
-            test_is_cmt_option()
             self.hit_thresholds = straxen.get_correction_from_cmt(self.run_id,
                 self.config['hit_min_amplitude_he'])
         # if hitfinder_thresholds config
