+++ straxen/plugins/pulse_processing.py
@@ -5,7 +5,7 @@ import numpy as np
 import strax
 import straxen
 
-from straxen.get_corrections import is_cmt_option, test_is_cmt_option
+from straxen.get_corrections import is_cmt_option
 
 export, __all__ = strax.exporter()
 __all__ += ['NO_PULSE_COUNTS']
@@ -161,7 +161,6 @@ class PulseProcessing(strax.Plugin):
         # Check config of `hit_min_amplitude` and define hit thresholds
         # if cmt config
         if is_cmt_option(self.config['hit_min_amplitude']):
-            test_is_cmt_option()
             self.hit_thresholds = straxen.get_correction_from_cmt(self.run_id,
                 self.config['hit_min_amplitude'])
         # if hitfinder_thresholds config
@@ -274,7 +273,6 @@ class PulseProcessingHighEnergy(PulseProcessing):
         # Check config of `hit_min_amplitude` and define hit thresholds
         # if cmt config
         if is_cmt_option(self.config['hit_min_amplitude_he']):
-            test_is_cmt_option()
             self.hit_thresholds = straxen.get_correction_from_cmt(self.run_id,
                 self.config['hit_min_amplitude_he'])
         # if hitfinder_thresholds config
