+++ straxen/plugins/veto_pulse_processing.py
@@ -3,7 +3,7 @@ import numpy as np
 import numba
 import straxen
 
-from straxen.get_corrections import is_cmt_option, test_is_cmt_option
+from straxen.get_corrections import is_cmt_option
 
 export, __all__ = strax.exporter()
 
@@ -67,7 +67,6 @@ class nVETOPulseProcessing(strax.Plugin):
         # Check config of `hit_min_amplitude_nv` and define hit thresholds
         # if cmt config
         if is_cmt_option(self.config['hit_min_amplitude_nv']):
-            test_is_cmt_option()
             self.hit_thresholds = straxen.get_correction_from_cmt(self.run_id,
                 self.config['hit_min_amplitude_nv'])
         # if hitfinder_thresholds config
@@ -211,7 +210,6 @@ class muVETOPulseProcessing(nVETOPulseProcessing):
         # Check config of `hit_min_amplitude_mv` and define hit thresholds
         # if cmt config
         if is_cmt_option(self.config['hit_min_amplitude_mv']):
-            test_is_cmt_option()
             self.hit_thresholds = straxen.get_correction_from_cmt(self.run_id,
                 self.config['hit_min_amplitude_mv'])
         # if hitfinder_thresholds config
